Refactor: ui 폴더 세분화
1. atomic한 폴더를 시도해보고 싶었습니다.
2.
- 원한 것: Card를 구성하는 컴포넌트들을 재사용할 수 있게 하고 싶었습니다.
- 문제사항: 카드를 구성하는 컴포넌트들을 atoms의 card라는 폴더에 넣고 molecules 폴더의 card 폴더에서 가져와서 다양한 카드의 이름으로 조직하여  소비하고 싶었습니다. 여기까지는 문제가 없었습니다. 문제는 `useCardProvider`가 같은 context를 받기 때문에 다양성을 열어두지 못한다는 점이었습니다. 그래서 molecules/card에 전부 몰아뒀습니다. 일단은 linkCard에서만 쓸 수 있도록 해놨습니다.
- 시도한 것: 처음에는 이 문제를 해결하기 위해 카드를 구성하는 컴포넌트들이 props로 useProvider라는 파라미터로 useProvider훅을 각각 제공받아서 제공받는 useProvider 별로 다른 동작을 하게하려고 구상했었습니다. 하지만 그렇게 될 경우 useProvider훅이 context로 갖게 되는 타입이 너무 다양해지고 복잡해지는 문제가 있습니다. 그렇게 되면 사용처에서 useProvider에 어떤 것이 들어 있을 지 장담하지 못하기 때문에 타입가드를 많이 해야 할 것 같습니다. 그래서 일단은 이렇게 구성만 해놓았습니다.
