+++ src/components/ui/molecules/card/linkCard/CardCover.jsx
